Turn Volvo spider multinational
